+++ src/main/java/subway/domain/SubwayManager.java
@@ -1,5 +1,11 @@
 package subway.domain;
 
+import subway.domain.line.Line;
+import subway.domain.line.LineManager;
+import subway.domain.line.LineRepository;
+import subway.domain.State;
+import subway.domain.section.SectionManager;
+import subway.domain.station.StationManager;
 import subway.view.InputView;
 import subway.view.OutputView;
