+++ src/main/java/subway/exception/NoSuchLineException.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.exception;
 
 public class NoSuchLineException extends RuntimeException{
     private static final String ERROR_MESSAGE = "[ERROR] 해당 노선은 존재하지 않습니다.";
