+++ src/main/java/subway/exception/NoSuchStationException.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.exception;
 
 public class NoSuchStationException extends RuntimeException{
     private static final String ERROR_MESSAGE = "[ERROR] 해당 역은 존재하지 않습니다.";
